requirements.txt
django==1.10
pillow==4.2.1
PyMySQL==0.8.0
djangorestframework-jwt==1.11.0
djangorestframework==3.7.7
Markdown==2.6.11
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-import-export==0.7.0
django-reversion==2.0.12
future==0.16.0
httplib2==0.9.2
six==1.11.0
requests==2.18.4
coreapi==2.3.3
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
raven